Introduce a dynamic addon imp registry to support Foreign Addon Imps
A Foreign Addon Imp is an addon implementation that is put outside of
gravyvalet codebase offered as a Django application. A collection of
Foreign Addon Imps is distributed as a Django project.

AddonRegistry class is the dynamic addon imp registry that holds the
addon imps in use. Even the built-in addon implementation needs to be
registered to be used, by adding the name and imp number pair to
`app.settings.ADDON_IMPS` configuration. `AddonImpNumbers` enum is
discarded. In addition, AddonImpRegistry class replaces the public
interface of `addon_service.common.known_imps` module.
Foreign Addon Imps needs to be registered to `ADDON_IMPS` too. Moreover,
they needs to be registered to `INSTALLED_APPS`.